move queries into separate files
