+++ packages/twenty-front/src/modules/activities/emails/queries/getTimelineMessagesFromCompanyId.ts
@@ -1,0 +1,15 @@
+import { gql } from '@apollo/client';
+
+export const getTimelineMessagesFromCompanyId = gql`
+  query GetTimelineMessagesFromCompanyId($companyId: String!) {
+    getTimelineMessagesFromCompanyId(companyId: $companyId) {
+      body
+      numberOfMessagesInThread
+      read
+      receivedAt
+      senderName
+      senderPictureUrl
+      subject
+    }
+  }
+`;
