+++ packages/twenty-front/src/modules/activities/emails/queries/getTimelineMessagesFromPersonId.ts
@@ -1,0 +1,15 @@
+import { gql } from '@apollo/client';
+
+export const getTimelineMessagesFromPersonId = gql`
+  query GetTimelineMessagesFromPersonId($personId: String!) {
+    getTimelineMessagesFromPersonId(personId: $personId) {
+      body
+      numberOfMessagesInThread
+      read
+      receivedAt
+      senderName
+      senderPictureUrl
+      subject
+    }
+  }
+`;
